Pin Alpine version and force clean build context
There are some OLCI image build applications like `buildah` and
`podman` that do not respect `.dockerignore`. If the build context is
initialized with a locally built `./artifacts` folder, there might be
a libc / musl mismatch, and the integration tests will fail with `file
not found`.

This commit ensures the build context is cleaned on every image build,
ensuring the binaries are rebuilt, and built against the right dynamic
link. The commit also pins the Alpine image version, and thus pins the
musl version that both the build and runtime images will use.